Refactor functional tests for vaccine cards and COVID-19 tests AB#11871
