+++ Testing/functional/tests/cypress/fixtures/LaboratoryService/covidTest.json
@@ -14,7 +14,9 @@
                 "testStatus": "Corrected",
                 "testOutcome": "NoOrderFound",
                 "resultTitle": "No Test Found",
-                "resultDescription": "No Covid results were found. Please confirm that the information you provided is correct.</p><p> It has been more than 48 hours since the date of testing provided. Please contact your primary care provider if you still have symptoms.",
+                "resultDescription": [
+                    "No Covid results were found. Please confirm that the information you provided is correct.</p><p> It has been more than 48 hours since the date of testing provided. Please contact your primary care provider if you still have symptoms."
+                ],
                 "resultLink": null
             }
         ]
